% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependent_features_fd.R
\name{add_dependent_feature_fd}
\alias{add_dependent_feature_fd}
\title{Add Dependent Features Football Data}
\usage{
add_dependent_feature_fd(x, feature)
}
\arguments{
\item{x}{results database}

\item{feature}{see the function add_dependent_features_fd for a list of 
possible string inputs. Incorrectly named inputs will warn and but not stop.}
}
\description{
Add time independent features to football data results dataset. 
This is a convienient wrapper to add all the features in a single line.
}
