% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_empty_stadiums}
\alias{add_empty_stadiums}
\title{Add COVID empty stadiums}
\usage{
add_empty_stadiums(
  results,
  metadata_competitions = load_metadata()$competitions
)
}
\arguments{
\item{results}{results tibble}
}
\value{
results tibble
}
\description{
Add TRUE or FALSE if stadiums are empty, mainly due to COVID, as this will change the home advantage
dynamic
}
\note{
dates are taken from metadata to determine if stadiums are empty or partially empty. Through time I will add
functionality to note if certain matches were played behind closed doors
}
