% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_independent_features_fd.R
\name{add_independent_feature_fd}
\alias{add_independent_feature_fd}
\title{Add Independent Feature Football Data}
\usage{
add_independent_feature_fd(x, feature)
}
\arguments{
\item{x}{results dataset from football data which with column names cleaned}

\item{feature}{see the function add_independent_features_fd for a list of 
possible string inputs. Incorrectly named inputs will warn and but not stop.}
}
\description{
Add time independent features to football data results dataset.
}
\note{
Time independent features are features which can be derived from an 
observation (i.e. that match). Some features which could be useful are time
dependent e.g. average win % for past n matches depends on the teams past n
matches.
}
\examples{
See the function clean_results_fd for an implementation
}
