% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_independent_features_fd.R
\name{add_independent_features_fd}
\alias{add_independent_features_fd}
\title{Add Independent Features Football Data}
\usage{
add_independent_features_fd(x)
}
\arguments{
\item{x}{results dataset from football data which with column names cleaned,}
}
\description{
Add time independent features to football data results dataset. 
This is a convienient wrapper to add all the features in a single line.
}
\note{
Time independent features are features which can be derived from an 
observation (i.e. that match). Some features which could be useful are time
dependent e.g. average win % for past n matches depends on the teams past n
matches.
}
\examples{
See the function clean_results_fd for an implementation
}
