% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependent_features_fd.R
\name{add_match_numbers_season}
\alias{add_match_numbers_season}
\title{Add Match Numbers Season}
\usage{
add_match_numbers_season(x)
}
\arguments{
\item{x}{results database}
}
\description{
Adds in the match number in the season for 
the home team and the away team to the results database. For consistent 
syntax for adding features this function is a helper called by 
add_dependent_feature
}
