% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_matches_played.R
\name{add_matches_played}
\alias{add_matches_played}
\title{Add Matches Played}
\usage{
add_matches_played(results)
}
\arguments{
\item{results}{standard results tibble}
}
\value{
results tibble with home_matches_played_season and away_matches_played_season
}
\description{
Add matches played that season for each team
}
