% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_table_results_main.R
\name{add_missing_col_name}
\alias{add_missing_col_name}
\title{Add Missing Column Name}
\usage{
add_missing_col_name(x, col_name, value = NA_real_)
}
\arguments{
\item{x}{the tibble}

\item{col_names_starts_with}{the columns to extract with starts_with e.g. "home_odds"}
}
\description{
Helper Function. Checkes if a column is missing in a tibble x and if it is it adds the column in
}
