% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependent_features_fd.R
\name{add_promotion_relegation}
\alias{add_promotion_relegation}
\title{Add Promotion Relegation}
\usage{
add_promotion_relegation(.competition_id, database_results, .metadata)
}
\arguments{
\item{x}{results database}
}
\description{
Adds in indicator variables for teams newly promoted or 
relegated to a league for that season.
}
