% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependent_features_fd.R
\name{add_spell}
\alias{add_spell}
\title{Add Spell}
\usage{
add_spell(x)
}
\arguments{
\item{x}{results database}
}
\description{
Adds in which spell, since the start of the dataset, the teams
are currently in
}
