% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_sets.R
\name{allocate_sets}
\alias{allocate_sets}
\title{Allocate Sets}
\usage{
allocate_sets(results, set_sizes = c(set1 = 0.1, set2 = 0.2, set3 = 0.2,
  set4 = 0.5))
}
\arguments{
\item{results}{results database, likely filtered to only contain a single
competition_id}

\item{set_sizes}{a vector of decimals, each of which corresponds to the 
proportions of the rows of results to be allocated to that set. The names of
each element if included.}
}
\value{
results with a column named set
}
\description{
Create a variable which splits a data frame or tibble into different sets 
based on a supplied vector of probabilties which correspond to the size of 
each set
}
