% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get_metadata.R
\name{as_tibble_metadata_competition}
\alias{as_tibble_metadata_competition}
\title{As Tibble Metadata Competition}
\usage{
as_tibble_metadata_competition(metadata_competition)
}
\arguments{
\item{metadata_competition}{list element of metadata}
}
\description{
Helper - turns a competition element from metadata into a tibble
which can then be mapped into a tibble with map_dfr
}
\examples{
metadata_tibble <- get_metadata()[[1]] \%>\% as_tibble_metadata()
}
