% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bets.R
\name{build_bet_placement_matrix}
\alias{build_bet_placement_matrix}
\title{Build Bet Placement Matrix}
\usage{
build_bet_placement_matrix(probs, odds, min_advantage, max_odds)
}
\arguments{
\item{probs}{Tibble, data.frame or matrix of estimated match probabilities. Columns are outcomes, rows are matches.}

\item{odds}{Tibble, data.frame or matrix of bookmakers odds. Columns are outcomes, rows are matches.}

\item{min_advantage}{Minimum advantage needed before a bet is placed. Default: 0.1.}

\item{max_odds}{Maximum odds which a bet is placed it. You might limit odds to lower variance. Default: 5.}
}
\value{
A matrix the same size as the probs and odds matrices, composed of 1s and 0s indicating where a bet has been 
placed
}
\description{
Build a bet placement matrix
}
