% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_model.R
\name{build_ensemble}
\alias{build_ensemble}
\title{Build Ensemble}
\usage{
build_ensemble(predictions, ensemble_weights)
}
\arguments{
\item{predictions}{a list containing match probabilities, each column is an outcome, each row a match. Each list
element must be the same structure}

\item{ensemble_weights}{a vector of weights}
}
\value{
a matrix of weighted probabilities
}
\description{
Combine a list of predictions using given weights
}
