% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bets.R
\name{build_indicator_matrix}
\alias{build_indicator_matrix}
\title{Build Indicator Matrix}
\usage{
build_indicator_matrix(x)
}
\arguments{
\item{x}{vector of outcomes, must be a factor}
}
\value{
An indicator matrix
}
\description{
Build an indicator matrix of 1s and 0s based on a vector of outcomes
}
\details{
Each column is a level from x. A 1 indicates the event occurred, a 0 it didn't.
}
