% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_data_poisson.R
\name{build_model_data_poisson}
\alias{build_model_data_poisson}
\title{Build model data poisson}
\usage{
build_model_data_poisson(results, x1, x2, current_date, xi = 0.0016)
}
\arguments{
\item{results}{results database}

\item{x1}{home goals, uses tidy eval so can specify home_goals_half1 etc}

\item{x2}{away goals, uses tidy eval so can specify away_goals_half1 etc}

\item{current_date}{what date to calculate the exponential time weight from}

\item{xi}{the time weight parameter}
}
\description{
Reformat the results data into an appropriate format for 
modelling using the basic, time-weighted Poisson model
}
