% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_advantage.R
\name{calc_advantage}
\alias{calc_advantage}
\title{Calculate Advantage}
\usage{
calc_advantage(probs, odds)
}
\arguments{
\item{probs}{true / estimated probabilities}

\item{odds}{odds on offer at the bookmaker}
}
\value{
the advantage as a percentage
}
\description{
Calculate the percentage advantage for a given bet
}
\details{
Advantage is equal to expected profit
}
\examples{
 modelled_probs <- 0.5
 bookmakers_odds <- 2.1

calc_advantage(modelled_probs, bookmakers_odds)

# Alternatively you could measure your tipped odds against closing odds

pinnacle_closing_odds <- 2.05

calc_advantage(1 / bookmakers_odds, pinnacle_closing_odds)

}
