% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_competition_elo.R
\name{calc_competition_elo}
\alias{calc_competition_elo}
\title{Calculate Competition ELO}
\usage{
calc_competition_elo(.competition_id, results, message_user = FALSE)
}
\arguments{
\item{.competition_id}{see get_metadata for format of competition ids}

\item{results}{standard results tibble but with 3 columns for home_prob,
draw_prob and away_prob. These probabilities should be implied probabilities
derived from bookmakers odds.}
}
\description{
Calculate ELO ratings for a competition, over multiple seasons
if necessary
}
