% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{calc_ensemble_weights}
\alias{calc_ensemble_weights}
\title{Calculate Ensemble Weights}
\usage{
calc_ensemble_weights(predictions, observed)
}
\arguments{
\item{predictions}{see \link[build_ensemble]}

\item{observed}{actual outcomes, possibly closing implied probs from closing odds or 0 1 indicators for actual 
outcomes}
}
\value{
a vector of optimised weights
}
\description{
Optimises a set of weights given a list of separate match outcome probabilities and the observed probabilities
}
