% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_table_results_main.R
\name{calc_max_odds}
\alias{calc_max_odds}
\title{Calculate Max Odds}
\usage{
calc_max_odds(x, col_names_starts_with)
}
\arguments{
\item{x}{the tibble}

\item{col_names_starts_with}{the columns to extract with starts_with e.g. "home_odds"}
}
\value{
maximum odds vector
}
\description{
Helper Function. Calculate the maximum odds columns from columns specified by starts_with
}
