% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rps.R
\name{calc_rps}
\alias{calc_rps}
\title{Calculate Ranked Probability Score}
\usage{
calc_rps(predictions, observed)
}
\arguments{
\item{predictions}{matrix, data.frame or tibble of probabilties, columns are 
outcomes and rows are observations}

\item{observed}{same as predictions, typically a 1 indicates the outcome 
occured and 0 it didnt though you could use bookmakers implied probabilties}
}
\value{
vector of scores
}
\description{
Calculate the ranked probability score
}
