% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_season_elo.R
\name{calc_season_elo}
\alias{calc_season_elo}
\title{Calculate Season ELO}
\usage{
calc_season_elo(results, elo_record)
}
\arguments{
\item{results}{set of results to calculate ELO rating for}

\item{elo_record}{starting ELO ratings for each team in the results data}
}
\description{
Calculate ELO ratings for a single season
}
\note{

}
