% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependent_features_fd.R
\name{calc_spell}
\alias{calc_spell}
\title{Calculate Spell}
\usage{
calc_spell(x)
}
\arguments{
\item{.competition_id}{competition_id to construct spells dataset for}

\item{database_results}{results database}
}
\value{

}
\description{
Adds in which spell, since the start of the dataset, the teams are currently in
}
\note{
A helper function for add_spell, see above
}
