% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_time_weights_exp.R
\name{calc_time_weights_exp}
\alias{calc_time_weights_exp}
\title{Calculate Time Weights Exponential}
\usage{
calc_time_weights_exp(match_dates, current_date, xi = 0.0016)
}
\arguments{
\item{match_dates}{dates of football matches, date object}

\item{current_date}{what date to calculate the exponential time weight from}

\item{xi}{the time weight parameter}
}
\description{
Uses an exponetial time weighting for modelling, weighting more
recent matches more strongly. This is derived from the time weighting method
used by Dixon-Coles in their version of the double Poisson model
}
