% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sets.R
\name{check_teams}
\alias{check_teams}
\title{Check Teams}
\usage{
check_teams(test_set, training_set)
}
\arguments{
\item{test_set}{football results database for testing}

\item{training_set}{football results database for training}
}
\description{
Helper function that checks teams in test set exist in a 
training set and filters them out if not
}
