% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_odds_fd.R
\name{clean_odds_fd}
\alias{clean_odds_fd}
\title{Clean odds from football data}
\usage{
clean_odds_fd(file, data_type = "main_league")
}
\arguments{
\item{file}{name of file downloaded from football-data.co.uk}

\item{data_type}{one of "main_league", "extra_league". See get_metdata.}
}
\description{
Clean results data from www.football-data.co.uk main leagues and
extra leagues
}
\examples{
file <- "C:/Users/Neil/Documents/football-betting/data/raw/eng_pl_2018_2019.csv"
clean_odds_fd(file)
}
