% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_odds_fd.R
\name{clean_odds_fdm}
\alias{clean_odds_fdm}
\title{Clean odds from football data main leagues}
\usage{
clean_odds_fdm(file)
}
\arguments{
\item{file}{name of file downloaded from football-data.co.uk}
}
\description{
Clean results data from www.football-data.co.uk main leagues
}
\examples{
file <- "C:/Users/Neil/Documents/football-betting/data/raw/eng_pl_2018_2019.csv"
clean_odds_fdm(file)
}
