% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_results_fd.R
\name{clean_results_fd}
\alias{clean_results_fd}
\title{Clean results from football data}
\usage{
clean_results_fd(file, data_type = "main_league")
}
\arguments{
\item{file}{name of file downloaded from football-data.co.uk}

\item{data_type}{either "main_league" or "extra_league". See ?get_metadata
for more info on different types of league at www.football-data.co.uk.}
}
\description{
Clean results data from www.football-data.co.uk
}
\examples{
file <- "C:/Users/Neil/Documents/football-betting/data/raw/eng_pl_2018_2019"
clean_results(file)
}
