% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_access_dbi.R
\name{connect_to_access_dbi}
\alias{connect_to_access_dbi}
\title{Connect to Access DBI}
\usage{
connect_to_access_dbi(file)
}
\value{
connection to your access database
}
\description{
Connect to access via a DBI connector
}
