% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-create_db_table_results_main.R
\name{create_db_table_results_main}
\alias{create_db_table_results_main}
\title{Create Database Table Results Main}
\usage{
create_db_table_results_main(file, data_type = "fd_main")
}
\arguments{
\item{file}{filename of the raw football match data from football-data.co.uk}

\item{data_type}{one of fd_main or fd_extra depending on league type}
}
\value{
results main tibble ready for pushing to database
}
\description{
Creates the data frame in the format required by the results main table in the live football database
}
