% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_table_results_main.R
\name{create_db_table_results_main_fd_extra}
\alias{create_db_table_results_main_fd_extra}
\title{Create Database Table Results Main Football Data Extra Leagues}
\usage{
create_db_table_results_main_fd_extra(file)
}
\arguments{
\item{file}{filename of the raw football match data from football-data.co.uk}
}
\value{
results main tibble ready for pushing to database
}
\description{
Creates the data frame in the format required by the results main table in the live football database for extra 
leagues from football-data.co.uk
}
