% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bets.R
\name{create_dummy_sim_data}
\alias{create_dummy_sim_data}
\title{Create Dummy Simulation Data}
\usage{
create_dummy_sim_data(num_matches, num_outcomes, .seed = NA)
}
\arguments{
\item{num_matches}{number of matches to simulate}

\item{num_outcomes}{how many outcomes can you create}

\item{.seed}{}
}
\description{
Create fake data for betting simulation
}
