% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-create_match_id.R
\name{create_match_id}
\alias{create_match_id}
\title{Create Match Id}
\usage{
create_match_id(competition_id, match_date, home_team, away_team)
}
\arguments{
\item{competition_id}{unique competition id}

\item{match_date}{of class date}

\item{home_team}{name of home team}

\item{away_team}{name of away team}
}
\value{
the match id
}
\description{
Construct unique match id from other variables
}
