% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sets.R
\name{create_sets}
\alias{create_sets}
\title{Create Sets}
\usage{
create_sets(results_competition, ..., num_seasons_initial = 3,
  num_seasons_cv = 3, min_matches_prediction = 6)
}
\arguments{
\item{results_competition}{football results database filtered so only a 
single competition_id}

\item{...}{variables in results_competition to keep}

\item{num_seasons_initial}{number of seasons required for initial training
set}

\item{num_seasons_cv}{number of seasons required for a middle validation set}

\item{min_matches_prediction}{how many matches in a season must each team 
have played before the models prediction is included in profit testing}
}
\description{
Creates a list of test and progressive training sets for fitting
betting models, splitting the test sets by match_date
}
\note{
Some rules around the training set / test set stuff
Use 3 seasons worth of data to initiallise. 
Season 1 - only training set
Season 2 - only training set
Season 3 - only training set
}
\examples{
results \%>\%
  read_rds(paste0(wd$data, 'database_results.rds')) \%>\%
  filter(competition_id = "sco_champ") \%>\%
  create_sets()
}
