% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_competition_id.R
\name{error_check_competition_id}
\alias{error_check_competition_id}
\title{Error Check Competition ID}
\usage{
error_check_competition_id(competition_id)
}
\arguments{
\item{competition_id}{See \link[panenkar]{get_metadata} for format details}
}
\value{
\code{TRUE} or \code{FALSE} indicating validity of the format of
competition_id
}
\description{
Check that competition_id is in a correct format
}
\examples{

error_check_competition_id("england_pl") # Will return FALSE
error_check_competition_id("eng_pl") # TRUE
}
