% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_season_id.R
\name{error_check_season_id}
\alias{error_check_season_id}
\title{Error Check Season ID}
\usage{
error_check_season_id(season_id)
}
\arguments{
\item{season_id}{See \link[panenkar]{sequence_seasons} for format details}
}
\value{
\code{TRUE} or \code{FALSE} indicating validity of the format of
season_id
}
\description{
Check that competition_id is in a correct format
}
\examples{

.error_check_season_id("2018_2019") # Will return TRUE
.error_check_season_id("2018_2020") # FALSE
.error_check_season_id(2018) # TRUE
}
