% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_element.R
\name{extract_element}
\alias{extract_element}
\title{Extract Element}
\usage{
extract_element(x, rows, ..., unlisted = TRUE)
}
\arguments{
\item{x}{tibble or data.frame}

\item{rows}{vector of numbers specifying the row or rows of to be extracted}

\item{...}{column name/names in x to extract using tidy evaluation}
}
\value{

}
\description{
Extract an element/elements from a tibble/data.frame. Handy 
shortcut bolted on  dplyr manipulations when piping.
}
\examples{

library(dplyr)

tb <- tibble(indices = 1:10, rand_nums = runif(10), 
             rand_letters = sample(letters, 10))

tb \%>\% 
  extract_element(5, rand_nums)
}
