% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_latest_elo.R
\name{extract_latest_elo}
\alias{extract_latest_elo}
\title{Extract Latest ELO}
\usage{
extract_latest_elo(.competition_id, results_elo,
  latest_season_teams = TRUE, sorted = TRUE)
}
\arguments{
\item{results_elo}{results tibble with elo ratings included}

\item{latest_season_teams}{default is TRUE, if TRUE only teams in the latest
season are returned}

\item{sorted}{do you want to sort the elo ratings from largest to smallest?}

\item{.competition_d}{The competition id}
}
\description{
Given extract the latest elo ratings for a specified competition
}
