% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_teams_by_season.R
\name{extract_teams_by_season}
\alias{extract_teams_by_season}
\title{Extract Teams by Season}
\usage{
extract_teams_by_season(x)
}
\arguments{
\item{x}{results database filtered so only a single competition id is present}
}
\value{
tibble with details of teams and season_ids
}
\description{
Returns a tibble with details of teams which were in a 
particular league by season
}
