% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-season-ids.R
\name{f_season}
\alias{f_season}
\title{F Season}
\usage{
f_season(season_ids, remove_na, f)
}
\arguments{
\item{season_ids}{season_id in one of two formats, either a single 4 digit number representing the year or a 
character string in the format 1999_2000 to represent seasons which extend over successive years, typical in European
football.}

\item{remove_na}{should NAs be ignored, default TRUE}

\item{f}{the function to use on the season_ids, min or max}
}
\description{
Helper for min and max season functions. The code is the same expect the call to the min or max function so to avoid
duplication the functions min or max are passed to the f_season function.
}
