% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_results_fd.R
\docType{data}
\name{file}
\alias{file}
\title{Clean results from football data extra leagues}
\format{An object of class \code{character} of length 1.}
\usage{
file
}
\arguments{
\item{file}{name of file downloaded from football-data.co.uk}
}
\description{
Clean results data from www.football-data.co.uk extra leagues
}
\details{
Will change this to a helper function once more data sources added.
}
\examples{
file <- "C:/Users/Neil/Documents/football-betting/data/raw/eng_pl_2018_2019"
clean_results_fdm(file)
}
\keyword{datasets}
