% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_model.R
\name{gd_add_ratings}
\alias{gd_add_ratings}
\title{Goal Difference Model Add Ratings}
\usage{
gd_add_ratings(results, xi, max_gd = NA, min_matches = 10)
}
\arguments{
\item{results}{results data set}

\item{xi}{time weight parameters}

\item{max_gd}{maximum goal difference, Default: NA}

\item{min_matches}{minimum matches, Default: 10}
}
\value{
results with home_rating, away_rating and match_rating columns
}
\description{
Add match rating to results tibble
}
