% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_model.R
\name{gd_calc_rating}
\alias{gd_calc_rating}
\title{Goal Difference Model Calculate Rating}
\usage{
gd_calc_rating(results, .team, current_date, xi, max_gd = NA, min_matches = 10)
}
\arguments{
\item{historic_results}{results to be used for the fitting the model}

\item{fit}{fitted probit model using clm, either results ~ match_rating or over_under_2_5 (etc) ~ match_rating}

\item{match_rating}{defined as the home_rating - away_rating, where a teams rating is exp weighted average of 
previous matches goal difference}

\item{market}{see \code{\link{gd_simulate_matches}}}

\item{over_under_goals}{see \code{\link{gd_simulate_matches}}}
}
\value{
match probabilities
}
\description{
Predict matches using the basic goal difference rating model
}
