% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_model.R
\name{gd_optim_xi}
\alias{gd_optim_xi}
\title{Goal Difference Model Optimize Xi}
\usage{
gd_optim_xi(training_set, test_set, start_xi, max_gd, min_matches)
}
\arguments{
\item{training_set}{see \code{\link{gd_simulate_matches}}}

\item{test_set}{see \code{\link{gd_simulate_matches}}}

\item{start_xi}{starting value of xi to try}

\item{max_gd}{see \code{\link{gd_simulate_matches}}}

\item{min_matches}{see \code{\link{gd_simulate_matches}}}
}
\value{
list, optim output including value and optimal parameters
}
\description{
Optimize the xi time weight parameter for the goal difference model
}
