% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_model.R
\name{gd_simulate_matches}
\alias{gd_simulate_matches}
\title{Goal Difference Model Simulate Matches}
\usage{
gd_simulate_matches(
  training_set,
  test_set,
  xi,
  max_gd = NA,
  min_matches = 10,
  markets = c("result", "over_under"),
  over_under_goals = 2.5
)
}
\arguments{
\item{training_set, }{columns of which must be "home_team", "away_team", ("result" and/or "over_under_2_5" or similar
depending on the market specified) and "match_date"}

\item{test_set, }{same column restrictions as training_set, the set to get outcome probabilities for}

\item{xi}{the time weight parameter for weighting the historic goal difference to derive the rating. This is just 
using the same Poisson regression time weighting approach but xi may be different}

\item{max_gd}{maximum goal difference to consider, if > than this value it is trimmed, Default: NA results in no
trimming. Trimming may be helpful for removing outliers.}

\item{min_matches}{minimum number of matches a team must have played before a rating is assigned. Default: 10.}

\item{markets}{which markets to calculate probabilities for, Default: c("result", "over_under")}

\item{over_under_goals}{used if '"over_under" present in markets, Default: 2.5}

\item{bind_to_test}{if FALSE just the probabilties will be outputted. THis is good for optimising parameters, 
Default: TRUE}
}
\value{
the predicted probabilities possibly binded to the test set
}
\description{
Simulate matches using the basic goal difference rating model
}
