% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_odds_fd.R
\name{get_bookmaker_lookup_football_data}
\alias{get_bookmaker_lookup_football_data}
\title{Get bookmaker lookup table for football data bookmakers}
\usage{
get_bookmaker_lookup_football_data()
}
\description{
A quick way to retrieve a lookup table which translates the
names of the bookmakers in the www.football-data.co.uk datasets into a
cleaner format
}
