% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-get_from_football_data.R
\name{get_from_football_data}
\alias{get_from_football_data}
\title{Get From Football Data}
\usage{
get_from_football_data(competition_id, season_ids = NA, path)
}
\arguments{
\item{competition_id}{See \link[panenkar]{get_metadata} for format details}

\item{season_ids}{See \link[panenkar]{sequence_seasons} for format details}

\item{path}{Folder to save the downloaded data into}
}
\description{
Download results and odds data from www.football-data.co.uk
}
\details{
Will change this to a helper function once more data sources added.
}
\examples{
get_from_football_data("eng_pl")
}
