% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R
\name{is_matrix_df_tibble}
\alias{is_matrix_df_tibble}
\title{Is Matrix, Data Frame or Tibble}
\usage{
is_matrix_df_tibble(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE or FALSE
}
\description{
Tests if an object is an matrix, data.frame or tibble
}
