% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R
\name{is_na_inf_nan}
\alias{is_na_inf_nan}
\title{Is NA Inf NaN}
\usage{
is_na_inf_nan(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical vector or matrix, data frames and tibbles will be converted to matrices
}
\description{
Test if an object is NA, Inf, -Inf or NaN
}
