% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R
\name{is_plain_vector}
\alias{is_plain_vector}
\title{Is Plain Vector}
\usage{
is_plain_vector(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE or FALSE
}
\description{
Tests if an object is an atomic vector and not a matrix, like output from c()
}
