% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_same_size.R
\name{is_same_size}
\alias{is_same_size}
\title{Is Same Size}
\usage{
is_same_size(...)
}
\arguments{
\item{...}{}
}
\value{
TRUE or FALSE
}
\description{
Test if matrices, data.frames or tibbles, or a combination of these have the same number of rows and 
             columns
}
\details{
DETAILS
}
\examples{

probs <- matrix(runif(9), nrow = 3)
odds <- matrix(runif(9, min = 1, max = 25), nrow = 3)
outcomes <- factor(sample(c("home", "draw", "away"), size = 9, replace = TRUE), levels = c("home", "draw", "away"))

betting_stats <- simulate_bets(probs, odds, outcomes)

}
