% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_ids.R
\name{is_season_id}
\alias{is_season_id}
\title{Check Season ID}
\usage{
is_season_id(season_id)
}
\arguments{
\item{season_id}{the season_id to check}
}
\value{
a list, first element is TRUE or FALSE telling you if the argument was a season_is, the 2nd is the error
message
}
\description{
Check that season_id is in a correct format
}
\note{
this is intended as helper function for verifying arguments
}
