% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_clean_teams.R
\name{manual_clean_teams}
\alias{manual_clean_teams}
\title{Manual Clean Teams}
\usage{
manual_clean_teams(database)
}
\arguments{
\item{database}{either database_results or database_odds}
}
\description{
Manually correct some errors in football data. See team-notes.R
for details of errors
}
\examples{
manual_clean_teams(database_results)
}
