% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-season-ids.R
\name{min_season}
\alias{min_season}
\title{Minimum Season}
\usage{
min_season(season_ids, remove_na = TRUE)
}
\arguments{
\item{season_ids}{season_id in one of two formats, either a single 4 digit number representing the year or a 
character string in the format 1999_2000 to represent seasons which extend over successive years, typical in European
football.}

\item{remove_na}{should NAs be ignored, default TRUE}
}
\value{
The minimum season_id
}
\description{
Determine the minimum season from a vector of season ids
}
