% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_teams_by_season.R
\name{mutate_season_id}
\alias{mutate_season_id}
\title{Mutate Season ID}
\usage{
mutate_season_id(.x, .season_id)
}
\arguments{
\item{.x}{results database}

\item{.season_id}{season_id}
}
\value{
.x with season_id variable
}
\description{
Helper for add_promotion_relegation
}
\note{
Helper function for extract_teams_by_season
}
