% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-poisson.R
\name{optim_xi_poisson_gd}
\alias{optim_xi_poisson_gd}
\title{Optimise xi Poisson Goal Difference}
\usage{
optim_xi_poisson_gd(
  xi,
  results_estimate_xi,
  match_rows,
  goal_difference_actual
)
}
\arguments{
\item{xi}{the value of xi to use in the time weight}

\item{results_estimate_xi}{results dataset to estimate xi off of}

\item{match_rows}{rows in the results dataset to estimate xi from}

\item{goal_difference_actual}{home_goals - away_goals, what really happened to compute mean square error of predicted
on}
}
\value{
mean square error in actual and predicted goal difference
}
\description{
Optimise xi parameter of time weighting for poisson model using the difference in goal difference as the evaluation
metric
}
