% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-poisson.R
\name{poisson_predict_match}
\alias{poisson_predict_match}
\title{Poisson Predict Match}
\usage{
poisson_predict_match(
  home_team,
  away_team,
  fit,
  max_goals = 8,
  markets = c("result", "over_under", "both_teams_to_score"),
  over_under_goals = 2.5
)
}
\arguments{
\item{home_team}{the name of the home_team}

\item{away_team}{the name of the away_team}

\item{fit}{fitted glm using poisson regression}

\item{max_goals}{the maximum number of goals to consider in a match}

\item{markets, }{vector contain one or more of "result", "over_under" or "both_teams_to_score"}

\item{over_under_goals}{a number usual 0.5, 1.5, 2.5, etc though could be a whole number. Only comes into play if 
markets is "over_under"}
}
\value{
a tibble of probabilities for each outcome
}
\description{
Predict a match using poisson regression
}
