% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson_predict_matches}
\alias{poisson_predict_matches}
\title{Poisson Predict Matches}
\usage{
poisson_predict_matches(
  fixtures,
  fit,
  max_goals = 8,
  markets = c("result", "over_under", "both_teams_to_score"),
  over_under_goals = 2.5
)
}
\arguments{
\item{fixtures}{a tibble columns of which must be "home_team", "away_team", "home_goals", "away_goals"}

\item{fit}{the fitted poisson or zero-inflated poisson model}

\item{max_goals}{the maximum number of goals to consider in a match}

\item{markets, }{vector contain one or more of "result", "over_under" or "both_teams_to_score"}

\item{over_under_goals}{a number usual 0.5, 1.5, 2.5, etc though could be a whole number. Only comes into play if 
markets is "over_under"}
}
\value{
probabilities for outcomes of the various markets for each match in fixtures
}
\description{
Predict multiple matches using Poisson regression with no updating of training sets or anything
}
