% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_predict_matches_simple.R
\name{poisson_predict_matches_simple}
\alias{poisson_predict_matches_simple}
\title{Poisson Predict Matches Simple}
\usage{
poisson_predict_matches_simple(fixtures, fit, max_goals = 8,
  market = "result", over_under_goals = 2.5, zero_inflated = FALSE)
}
\arguments{
\item{fixtures}{match fixtures, columns of which must be "home_team" and "away_team". For the rest of the parameters
and return see poisson_predict_match.R}
}
\description{
Predict multiple matches using Poisson regression with no updating of training sets or anything
}
