% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_predict_matches_update.R
\name{poisson_predict_matches_update}
\alias{poisson_predict_matches_update}
\title{Poisson Predict Matches Update}
\usage{
poisson_predict_matches_update(training_set, test_set, xi = 0.0016,
  max_goals = 8, market = "result", over_under_goals = 2.5,
  zero_inflated = FALSE, weight_cut_off = NA,
  mark_promoted_relegated = TRUE)
}
\arguments{
\item{fixtures}{match fixtures, columns of which must be "home_team" and "away_team". For the rest of the parameters
and return see poisson_predict_match.R}
}
\description{
Predict multiple matches using Poisson regression, updating the training set in the process
}
