% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-poisson.R
\name{poisson_simulate_matches}
\alias{poisson_simulate_matches}
\title{Poisson Simulate Matches}
\usage{
poisson_simulate_matches(
  training_set,
  test_set,
  xi = 0.0016,
  max_goals = 8,
  markets = c("result", "over_under", "both_teams_to_score"),
  over_under_goals = 2.5,
  zero_inflated = FALSE,
  weight_cut_off = NA
)
}
\arguments{
\item{training_set, }{columns of which must be "home_team", "away_team", "home_goals", "away_goals", "match_date"}

\item{test_set, }{same column restrictions as training_set}

\item{xi}{parameter for the exponential time weighting, by default 0.0016}

\item{max_goals, }{maximum number of home goals and/or away goals to model}

\item{markets, }{vector contain one or more of "result", "over_under" or "both_teams_to_score"}

\item{over_under_goals}{typically n.5 where n in 1, 2, 3 .. etc, default = 2.5. Argument only used if "over_under"
listed in markets}

\item{zero_inflated}{default FALSE, if TRUE the zero inflated poisson model is used}

\item{weight_cut_off, }{NA by default can set a lower cut off e.g. 0.01 which will drop observations from model data
set to potentially improve speed with a potential accuracy trade off}
}
\value{
probabilities for outcomes of the various markets for each match in the test set
}
\description{
Predict multiple matches using Poisson regression, updating the training set in the process
}
