% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fixtures_elo.R
\name{predict_fixtures_elo}
\alias{predict_fixtures_elo}
\title{Predict Fixtures ELO}
\usage{
predict_fixtures_elo(.competition_id, results_elo, elo_latest)
}
\arguments{
\item{.competition_id}{see get_metadata for format of competition ids}

\item{results_elo}{standard results tibble but with columns for home_elo_pre
and away_elo_pre}
}
\description{
Uses logistic regression and historical elo ratings to predict
fair probabilities and odds for matches
}
