% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-season-ids.R
\name{previous_season}
\alias{previous_season}
\title{Previous Season}
\usage{
previous_season(season_id)
}
\arguments{
\item{season_id}{season_id in form XXXX_XXXX or XXXX}
}
\value{
season_id
}
\description{
Get previous season_id
}
