% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_build_model_data.R
\name{probit_build_model_data}
\alias{probit_build_model_data}
\title{Build model data probit}
\usage{
probit_build_model_data(results, x, current_date, xi = 0.0016)
}
\arguments{
\item{results}{results database}

\item{x}{result variable}

\item{current_date}{what date to calculate the exponential time weight from}

\item{xi}{the time weight parameter}
}
\description{
Reformat the results data into an appropriate format for modelling using probit model
}
