% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit_fit}
\alias{probit_fit}
\title{Probit Fit}
\usage{
probit_fit(model_data)
}
\arguments{
\item{model_data}{output from probit_build_model_data function}
}
\value{
fitted model
}
\description{
A wrapper for the polr function with the specs required for order probit model
}
