% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit_predict_match}
\alias{probit_predict_match}
\title{Probit Predict Match}
\usage{
probit_predict_match(home_team, away_team, fit, market = "result")
}
\arguments{
\item{home_team}{name of home team}

\item{away_team}{name of away team}

\item{fit}{fitted probit model, output from probit_fit}

\item{market}{market type only supports result currently Default: 'result'}
}
\value{
outcome probabilities
}
\description{
Predict a match using ordered probit model
}
