% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit_predict_matches}
\alias{probit_predict_matches}
\title{Probit Predict Matches Simple}
\usage{
probit_predict_matches(fixtures, fit, market = "result")
}
\arguments{
\item{fixtures}{match fixtures, columns of which must be "home_team" and "away_team". For the rest of the parameters
and return see probit_predict_match.R}
}
\description{
Predict multiple matches using probit model with no updating of training sets or anything
}
