% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit_simulate_matches}
\alias{probit_simulate_matches}
\title{Probit Predict Matches Update}
\usage{
probit_simulate_matches(training_set, test_set, market = "result")
}
\arguments{
\item{fixtures}{match fixtures, columns of which must be "home_team" and "away_team". For the rest of the parameters
and return see probit_predict_match.R}
}
\description{
Predict multiple matches using probit, updating the training set in the process
}
