% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_robust.R
\name{read_csv_robust}
\alias{read_csv_robust}
\title{Read CSV File with Check}
\usage{
read_csv_robust(...)
}
\arguments{
\item{...}{see arguments of dplyr::read_csv}
}
\value{
Outputs a list with elements named "csv", "error" and "error_message"
so that you can access the list to check if an error has occurred. If it has
occurred then you can throw a stop using the error_message to halt your code
otherwise extract the csv element, ideal for writing more robust code. This
also means you can suppress warnings around read_csv without fear of missing
something.
}
\description{
Read's in a CSV file with an error handler built in incase of
failure.
}
\examples{
file = "C:/Users/Neil/Documents/football-betting/data-raw/arg_pd_all.csv"

suppressWarnings(
  arg_pd <- read_csv_robust(file)
)

if(arg_pd$error == TRUE){
  stop(arg_pd$error_message)
} else {
  arg_pd <- arg_pd$csv
}
}
