% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_margin.R
\name{remove_margin}
\alias{remove_margin}
\title{Remove Margin}
\usage{
remove_margin(
  prob_with_margin,
  margin,
  num_selections = 3,
  method = "proportional"
)
}
\arguments{
\item{prob_with_margin}{the implied probability of a selection occurring. 
This is equal to 1 / bookmakers odds.}

\item{margin}{the bookmakers margin, this is equal to the sum of the implied
probability of all selections minus 1.}

\item{num_selections}{the number of selections, for the full time market 
this would be 3}

\item{method}{by default this is set to the 'proportional' method detailed
here: https://www.football-data.co.uk/The_Wisdom_of_the_Crowd_updated.pdf but
can also be the 'straight' method. The proportional method is thought to be
superior, straight method is just divide everything by the overround.}
}
\value{
The probability with the margin removed.
}
\description{
Given a bookmakers implied probability remove the bookmakers margin
}
\examples{
odds_home <- 2.5
odds_draw <- 3
odds_away <- 3.5

margin <- 1 / odds_home + 1 / odds_draw + 1/ odds_away
probs_margin_removed <- remove_margin(1/odds_home, margin)
}
