% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result}
\alias{result}
\title{Result}
\usage{
result(home_goals, away_goals)
}
\arguments{
\item{home_goals}{goals scored by the home team}

\item{away_goals}{goals scored by the away team}
}
\value{
The result, a character vector with possible values "home", "draw" or
"away"
}
\description{
Given a score, calculate the result
}
\examples{
result(2, 1)
}
