% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_number.R
\name{round_number}
\alias{round_number}
\title{Round Number}
\usage{
round_number(x, digits = 0)
}
\arguments{
\item{x}{number or numbers to be rounded}

\item{digits}{number of digits to round to, Default: 0}
}
\value{
rounded number
}
\description{
Robust rounding function which is same as round but always rounds up on n.5 no matter the value of n
}
\details{
R, and other languages do this weird thing where 1.5 is rounded to 2 an so is 2.5. Every .5 is rounded to 
the nearest even number. This isn't typically what you want in mathematics.
}
