% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_max_min.R
\name{row_numeric_function}
\alias{row_numeric_function}
\title{Row Numeric Function}
\usage{
row_numeric_function(
  x,
  ...,
  na_rm = TRUE,
  append_col = TRUE,
  new_col_name = NA,
  f
)
}
\arguments{
\item{x}{a tibble}

\item{...}{columns of the tibble to compute the rowwise maximum on, must be numeric}

\item{na_rm}{if TRUE ignore NAs if FALSE dont ignore}

\item{append_col}{if TRUE adds a column to the end of x with the maximum}

\item{new_col_name}{column name of the maximum, only used if append_col = TRUE, if left as name x_row_max used as
default column name}

\item{f}{the function name as string e.g. "max" for row maxes}
}
\value{
a vector
}
\description{
Apply the function f to a tibble x rowwise
}
\note{
in theory you could use a whole bunch of functions for f, have only written tests for max and min
}
