% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_seasons.R
\name{sequence_seasons}
\alias{sequence_seasons}
\title{Sequence Seasons}
\usage{
sequence_seasons(start_season, end_season)
}
\arguments{
\item{start_season}{Takes two formats of season, either a single 4 digit
number representing the year or a character string in the format 1999_2000 to
represent seasons which extend over successive years, typical in European
football}

\item{end_season}{End season in the same format as start_season. Must come
after start_season}
}
\value{
A sequence of seasons between and including the start and end
seasons.
}
\description{
Create a sequence of seasons between start and end seasons
}
\examples{
sequence_seasons("2015_2016", "2019_2020")
sequence_seasons(2000, 2008)
}
