% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bets.R
\name{simulate_bets}
\alias{simulate_bets}
\title{Simulate Bets}
\usage{
simulate_bets(
  probs,
  odds,
  outcomes,
  closing_odds,
  min_advantage = 0.05,
  start_bank = 100,
  stake = 1,
  max_odds = 5,
  tolerance_digits = 3
)
}
\arguments{
\item{probs}{Tibble, data.frame or matrix of estimated match probabilities. Columns are outcomes, rows are matches.}

\item{odds}{Tibble, data.frame or matrix of bookmakers odds. Columns are outcomes, rows are matches.}

\item{outcomes}{Factor, the levels of which correspond to a column in probs and odds. They must all be ordered in the
same way.}

\item{closing_odds}{Tibble, data.frame or matrix of bookmakers odds. These are the odds at close and not necessarily
bet on. These are used to check the closing line value (CLV) of the bets made.}

\item{min_advantage}{Minimum advantage needed before a bet is placed. Default: 0.1.}

\item{start_bank}{Starting bankroll units. Default: 100.}

\item{stake}{Stake to place per bet. Default: 1.}

\item{max_odds}{Maximum odds which a bet is placed it. You might limit odds to lower variance. Default: 5.}

\item{tolerance_digits}{Probabilities for a given match must sum to 1. However, some tolerance in this may be needed
depending on how they were calculated. Default: 3. This means a matches probabilities would 
not be flagged with they sum to 1+-0.001.}
}
\value{
A list of betting statistics.
}
\description{
Simulate advantage bets
}
\details{
DETAILS
}
\examples{

probs <- matrix(runif(9), nrow = 3)
odds <- matrix(runif(9, min = 1, max = 25), nrow = 3)
outcomes <- factor(sample(c("home", "draw", "away"), size = 9, replace = TRUE), levels = c("home", "draw", "away"))

betting_stats <- simulate_bets(probs, odds, outcomes)

}
