% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_elo.R
\name{update_elo}
\alias{update_elo}
\title{Update ELO}
\usage{
update_elo(elo_home, elo_away, home, draw, away, advantage_home = 80,
  x = 10, y = 400, z = 175, output_format = "tibble")
}
\arguments{
\item{elo_home}{The current elo rating of the home team}

\item{elo_away}{The current elo rating of the away team}

\item{away}{home, draw and away take one of the following formats. Either an
            indicator to show what the actual result was e.g. home = 1, 
            draw = 0, away = 0 for a home win or the probability of each 
            outcome derived via bookmakers odds or some other method. 
for more info on different types of league at www.football-data.co.uk.}
}
\description{
Update ELO rating
}
\note{
The functions default values are configured for the 'odds' methods 
      where probabilities derived via bookmakers odds are used. It has been
      shown that this method is superior for generating an ELO rating than
      standard methods of using outcomes only.
}
\examples{
odds <- c(2.5,3.2,4)
probs <- 1 / odds
probs <- probs / sum(probs)
elo_home <- 800
elo_away <- 950

update_elo(800, 950, probs[1], probs[2], probs[3], 
           output_format = "named_vector")
}
