% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_matches_played.R
\name{add_matches_played}
\alias{add_matches_played}
\title{Add Matches Played}
\usage{
add_matches_played(x)
}
\arguments{
\item{x}{results tibble with cols 'home_team', 'away_team' and 'season_id'}
}
\value{

}
\description{
Add matches played by each team for each season
}
