% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependent_features_fd.R
\name{add_spell}
\alias{add_spell}
\title{Add Spell}
\usage{
add_spell(.competition_id, database_results)
}
\arguments{
\item{.competition_id}{competition_id to construct spells dataset for}

\item{database_results}{results database}
}
\value{

}
\description{
Adds in which spell, since the start of the dataset, the teams are currently in
}
\note{
A helper function for add_dependent_feature, see above
}
