% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rps.R
\name{calc_rps}
\alias{calc_rps}
\title{Calculate Rank Probability Score}
\usage{
calc_rps(predicted, observed, check_predicted_row_sums = TRUE)
}
\arguments{
\item{predicted}{a matrix, data frame or tibble of probabilities, each row is a match and each column is an outcome
e.g. home, draw, away. Columns must sum to 1.}

\item{observed}{a vector of actual outcomes corresponding to the columns in predicted e.g if 1 then the outcome in
the first column of predicted actually occurred, you could also pass implied probabilities}

\item{check_predicted_row_sums}{if TRUE this will check that row sums in predicted are 1 at a tolerance level of 
0.0001.}
}
\value{
the rps for each match or average rps score. Smaller score is better.
}
\description{
Calculates the rank probability score of forecasts
}
