% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rps.R
\name{calc_rps}
\alias{calc_rps}
\title{Calculate Rank Probability Score}
\usage{
calc_rps(predicted, observed, average_all = FALSE)
}
\arguments{
\item{predicted}{a matrix, data frame or tibble of probabilities, each row is a match and each column is an outcome
e.g. home, draw, away. Columns must sum to 1.}

\item{observed}{a vector of actual outcomes corresponding to the columns in predicted e.g if 1 then the outcome in
the first column of predicted actually occurred}

\item{average_all}{default FALSE is TRUE the usual output, a vector of scores for each match, is averaged so only a
single value is returned.}
}
\description{
Calculates the rank probability score of forecasts
}
