% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{optim_ensemble}
\alias{optim_ensemble}
\title{Optimize Ensemble}
\usage{
optim_ensemble(ensemble_weights, predictions, observed)
}
\arguments{
\item{ensemble_weights}{see \code{\link{build_ensemble}}}

\item{predictions}{see \code{\link{build_ensemble}}}

\item{observed}{see \code{\link{calc_ensemble_weights}}}
}
\value{
rps for that weighting
}
\description{
Helper function for deriviing optimum model weights
}
