% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson_build_model_data}
\alias{poisson_build_model_data}
\title{Build model data poisson}
\usage{
poisson_build_model_data(
  results,
  x1,
  x2,
  current_date,
  xi = 0.0016,
  weight_cut_off = NA
)
}
\arguments{
\item{results}{results database}

\item{x1}{home goals, uses tidy eval so can specify half time goals if you like}

\item{x2}{away goals, uses tidy eval so can specify half time goals if you like}

\item{current_date}{what date to calculate the exponential time weight from}

\item{xi}{the time weight parameter}

\item{weight_cut_off}{if a weight is less than this cut off that row is dropped from the model data for efficiency
purposes}
}
\value{
formatted tibble ready for using poisson_fit
}
\description{
Reformat the results data into an appropriate format for modelling using the basic, time-weighted Poisson model
}
