% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-poisson.R
\name{poisson_fit}
\alias{poisson_fit}
\title{Poisson Fit}
\usage{
poisson_fit(model_data, zero_inflated = FALSE)
}
\arguments{
\item{model_data}{result data formated by poisson_build_model_data function}

\item{zero_inflated}{default FALSE, if TRUE will use zero inflated poisson}
}
\value{
fitted model
}
\description{
A wrapper for the glm function with the specs required for poisson regression model and similar for the zero inflated
poisson model
}
