% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_model.R
\name{probit_build_model_data}
\alias{probit_build_model_data}
\title{Probit Build Model Data}
\usage{
probit_build_model_data(results, x, current_date, xi = 0.0016)
}
\arguments{
\item{results}{results tibble typically from the live database table results-main}

\item{x}{variable where the results are, usually 'result' or maybe a half time or second half result variable}

\item{current_date}{base data for calculating time weights}

\item{xi}{base parameter in time weight description, Default: 0.0016}
}
\value{
model data tibble ready for input into probit_fit
}
\description{
Build model data for probit models based on full time result or goal difference
}
