% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_model.R
\name{probit_fit}
\alias{probit_fit}
\title{Probit Fit}
\usage{
probit_fit(model_data)
}
\arguments{
\item{model_data}{Output from \code{\link{probit_build_model_data}}}
}
\value{
fitted model
}
\description{
Fitted an ordered probit model
}
