% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_predict_match.R
\name{probit_predict_match}
\alias{probit_predict_match}
\title{Probit Predict Match}
\usage{
probit_predict_match(home_team, away_team, fit, market = "result")
}
\arguments{
\item{home_team}{the name of the home_team}

\item{away_team}{the name of the away_team}

\item{fit}{fitted model using ordered probit}

\item{market}{only "result" applicable currently}
}
\value{
a tibble of probabilities for each outcome
}
\description{
Predict a match using ordered probit model
}
