% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_model.R
\name{probit_simulate_matches}
\alias{probit_simulate_matches}
\title{Probit Simulate Matches}
\usage{
probit_simulate_matches(training_set, test_set, market = "result")
}
\arguments{
\item{training_set, }{columns of which must be "home_team", "away_team", "result", "match_date"}

\item{test_set, }{same column restrictions as training_set}

\item{market}{Only 'result' currently valid, argument present for future development Default: 'result'}
}
\value{
probabilities for outcomes  for each match in the test set
}
\description{
#' Predict multiple matches using ordered probit model, updating the training set in the process
}
