% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_margin.R
\name{remove_margin}
\alias{remove_margin}
\title{Remove Margin}
\usage{
remove_margin(x, method = "proportional")
}
\arguments{
\item{x}{a vector, matrix, data.frame or tibble of decimal odds}

\item{method}{one of 'proportional' or 'straight'. Default: 'proportional'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Remove the bookmakers margin from a set of odds
}
\details{
the 'proportional' method detailed here: https://www.football-data.co.uk/The_Wisdom_of_the_Crowd_updated.pdf 
but can also be the 'straight' method. The  proportional method is thought to be superior, straight method is just 
divide everything by the over round.
}
