% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_ids.R
\name{sequence_seasons}
\alias{sequence_seasons}
\title{Sequence Seasons}
\usage{
sequence_seasons(start_season, end_season)
}
\arguments{
\item{start_season}{a season_id}

\item{end_season}{a season_id must come after start_season}
}
\value{
A sequence of seasons between and including the start and end seasons.
}
\description{
Create a sequence of seasons between start and end seasons
}
\examples{
sequence_seasons("2015_2016", "2019_2020")
sequence_seasons(2000, 2008)
}
